refactor: #12 반복되는 코드 제거
