+++ app/account/setting/index.tsx
@@ -13,8 +13,15 @@ import { useRouter } from 'expo-router';
 import { Text, TouchableOpacity } from 'react-native';
 import styled from 'styled-components/native';
 
+const settingItems = [
+  { text: '알림 설정', route: '/account/setting/notifications' },
+  { text: '의견 남기기', route: '/account/setting/help' },
+  { text: '계정 관리', route: '/account/setting/account-status' },
+] as const;
+
 export default function SettingPage() {
   const router = useRouter();
+
   return (
     <Container>
       <Header
@@ -37,24 +44,13 @@ export default function SettingPage() {
         }
       />
       <MarginContainer>
-        <MyPageButton
-          text="알림 설정"
-          onPress={() => {
-            router.push('/account/setting/notifications');
-          }}
-        />
-        <MyPageButton
-          text="의견 남기기"
-          onPress={() => {
-            router.push('/account/setting/help');
-          }}
-        />
-        <MyPageButton
-          text="계정 관리"
-          onPress={() => {
-            router.push('/account/setting/account-status');
-          }}
-        />
+        {settingItems.map((item) => (
+          <MyPageButton
+            key={item.route}
+            text={item.text}
+            onPress={() => router.push(item.route)}
+          />
+        ))}
       </MarginContainer>
     </Container>
   );
@@ -65,6 +61,7 @@ const Container = styled.View`
   background-color: ${colors.bg[2]};
   padding: ${28 * height}px ${20 * width}px;
 `;
+
 const MarginContainer = styled.View`
   margin-top: ${12 * height}px;
   gap: ${8 * height}px;
